alii/requirements.txt
MySQL-python==1.2.5
pandas==2.0.0
python-dotenv==1.0.0
Requests==2.28.2
